Split: apertium-pretransfer.cc to pretransfer.h & pretransfer.cc
